classes/class06_decision/pyproject.toml
[project]
name = "ai-agent-lab-class06"
version = "0.1.0"
description = "Class06 - Decision flow with RAG score (LangGraph + FAISS)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.1.0,<1.2.0",
    "langchain-openai>=1.1.3,<1.2.0",
    "langgraph>=1.0.5,<1.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "langchain-community>=0.3.0",
    "faiss-cpu>=1.8.0",  
]


[tool.uv]
package = true